feat : put structure for HomeViewModel

+++ viewmodel/src/main/java/com/cairosquad/viewmodel/home/HomeEffect.kt
@@ -3,9 +3,9 @@ package com.cairosquad.viewmodel.home
 sealed class HomeEffect {
     data object NavigateToProfile: HomeEffect()
     data class NavigateToCategory(val category:String): HomeEffect()
-    data class NavigateToFeaturedMovie(val movieID :Int) : HomeEffect()
-    data class NavigateMovie(val movieID :Int) : HomeEffect()
-    data class NavigateSeries(val movieID :Int) : HomeEffect()
+    data class NavigateToFeaturedMovie(val movieID :Long) : HomeEffect()
+    data class NavigateMovie(val movieID :Long) : HomeEffect()
+    data class NavigateSeries(val movieID :Long) : HomeEffect()
     data object NavigateToSeeAllTopRated: HomeEffect()
     data object NavigateToSeeAllTrending: HomeEffect()
     data object NavigateToSeeAllFreeToWatch: HomeEffect()
